Redo diagnostic handling
Stop trying to attach diagnostics on `Scope` nodes coming from slang --
that has turned out to be very error prone when combined with the way
`Compilation` collects the diagnostics, and combined with the fact that
we don't want to operate on ASTs for which slang has issued errors.

Instead rely on our own diagnostic collection bypassing the
`Compilation` class. This has been made easier by recent `Driver`
changes.