Detect string parameters for RTLIL benefit
RTLIL has a notion of string constants. Although per the Verilog
standard string literals don't have a type which would distinguish them
from integral values, we can look at the syntax tree and guess when
a parameter looks to be taking strings. Add this detection for increased
compatibility with the Yosys native frontend.